chore(modularizeDiscoveryDetailsComponents): Initial commit, after creating new branch to avoid merge issues
