+++ src/Discovery/DiscoveryDetails/Components/FieldGrouping/TabField/DataDownloadList/DataDownloadList.tsx
@@ -5,8 +5,8 @@ import DataDownloadListItem from './Interfaces/DataDownloadListItem';
 import './DataDownloadList.css';
 import ProcessData from './Utils/ProcessData';
 import ActionButtons from './ActionButtons/ActionButtons';
-import { DiscoveryConfig } from '../../DiscoveryConfig';
-import { AccessLevel, DiscoveryResource, accessibleFieldName } from '../../Discovery';
+import { DiscoveryConfig } from '../../../../../DiscoveryConfig';
+import { AccessLevel, DiscoveryResource, accessibleFieldName } from '../../../../../Discovery';
 import { INITIAL_DOWNLOAD_STATUS } from './ActionButtons/DownloadUtils/Constants';
 import DownloadModal from './ActionButtons/DownloadModal/DownloadModal';
 import StandaloneDataDownloadButton from './ActionButtons/StandaloneDataDownloadButton';
