+++ src/Discovery/DiscoveryDetails/Components/FieldGrouping/TabField/DataDownloadList/Interfaces/DownloadStatus.ts
